staging: iio: adc: ad7192: Remove header file
These modification are done in order to get the driver out of staging. The
header file has been removed and additional fields have been included
inside the state structure, and are being used to configure the device. A
clock has been added to select between internal or external clock. Platform
data is no longer used for this driver.

Signed-off-by: Mircea Caprioru <mircea.caprioru@analog.com>